requirements.txt
langchain==1.2.15
langchain-community==0.3.31
langchain-core==1.2.26

langchain-text-splitters
langchain-huggingface
langchain-chroma
langchain-groq
langchain-openai

chromadb
sentence-transformers
unstructured[pdf]
nltk
pypdf
python-dotenv

flask
streamlit

langchain-pinecone

-e .